follow conventions for test examples.
